Added different voice

+++ events/discord/messageCreate.js
@@ -27,8 +27,8 @@ module.exports = {
                     return message.reply('No estás en un canal de voz').then(msg => {
                         setTimeout(() => msg.delete(), 15000)
                     });
-                if (command.inVoice && message.member.voice.channel.id !== message.guild.members.me.voice.channel.id) {
-                    
+                if (command.inVoice && message.member.voice.channel && message.guild.members.me.voice.channel && message.member.voice.channel.id !== message.guild.members.me.voice.channel.id) {
+                    console.log(message.guild.members.me.voice.channel);
                     return message.reply('No estás en el mismo canal de voz').then(msg => {
                         setTimeout(() => msg.delete(), 15000)
                     });
